[PLAT-13188] Migrate release metadata to "Release" implementation
Summary:
Create a repeatable migration that will move all ReleaseMetadata and
create Releases and Release Artifacts as needed. This is repeatable,
as the feature flag may be toggled off for some time, and will need to
perform the migration as new releases are added, until the flag
is toggled on.

Also added a quick fix where we could not update a release that had
multiple artifacts via the update api.

Test Plan:
created releases with "release metadata", enabled the release redesign and
restarted YBA.
The migration ran and imported all release metadata -> release.

Reviewers: muthu, anijhawan

Reviewed By: muthu

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D33482