[BACKPORT 2024.1] [PLAT-13256] Add sleep before removing tservers from blacklist.
Summary:
Add sleep before removing tservers from blacklist.
It seems there is a race condition where master may sometimes assign
tablets to tservers in prevoter state. This can cause incorrect
tablet balancing to happen in a cluster.
This imbalance is seen to persist for a long time.

This change introduces an optional wait parameter and some scafolding code
to add a 90second delay after deletion of the tserver to remove it from
the blacklist.

We chose 90s as the tserver failed heartbeat detection timeout is 60s.

Test Plan: phabricator-itest

Reviewers: sanketh, vkumar, sneelakantan

Reviewed By: vkumar

Subscribers: yugaware, nsingh

Tags: #jenkins-ready

Differential Revision: https://phorge.dev.yugabyte.com/D33524