[#11863] Fix flakiness in TestPgTransparentRestarts.selectStarLong* tests
Summary:
There is an assertion in TestPgTransparentRestarts tests for SELECTs
in parallel with INSERTs that checks for two SELECT statements
in a read committed transaction block to have different values atleast
once in the test run. However, with the current test run time, it was
possible by chance for some runs to not see such a case. This diff
removes that assertion.

Test Plan: Jenkins: test regex: .*TestPgTransparentRestarts.*

Reviewers: bkolagani

Reviewed By: bkolagani

Subscribers: yql

Differential Revision: https://phorge.dev.yugabyte.com/D33529