Merged PR 1306867: Flaky test fix in $currentDate update op
### Does this PR have any customer impact?
No
### Type (Feature, Refactoring, Bugfix, DevOps, Testing, Perf, etc)
Bugfix
### Does it involve schema level changes? (Table, Column, Index, UDF, etc level changes)
No
### Release Note (Refer [Template](../docs/release_notes/pgmongo_template.md))

### Description

Flaky test fix for $currentDate update op.
The currentDate related test in `bson_update_document_tests.sql` is flaly.
The test is taking a begin timestamp, then doing the update with $currentDate, then taking an end timestamp after the update, and expecting that the updated timestamp should be between the begin and end timestamps. But in the flaky case, it wasn't.

The error log:
`Updated timestamp "a" : 1710989509000 should be between 1710989507998 and 1710989508001`

Notice the updated timestamp for field "a" is exactly 1 second more than the begin/end timestamp.
It looks like, it was happening due to a rounding issue at the edge case (and which is why its flaky).