[#21631] DocDB: Avoid including hidden tables in generate_snapshot as of time
Summary:
Currently, Generate snapshot as of time is returning the set of tables and tablets that are in a `running` state as of `restore_time`. However, when we have a snapshot schedule, deleted tables are still marked as `RUNNING` but have a hidden flag set to true. These hidden tables should not be included in the output of generate snapshot. Otherwise, cloning will fail in case of a recreated table.
The diff fixes the issue by skipping to include the tables that are marked as hidden (eventhough they are in running state). It also skips the tablets that are hidden due to tablet splitting. This solves cloning issues: GH-21602 and GH-21577
Jira: DB-10526

Test Plan:
./yb_build.sh --cxx-test minicluster-snapshot-test --gtest_filter Colocation/MasterExportSnapshotTest.ExportSnapshotAsOfTimeWithHiddenTables/0
./yb_build.sh --cxx-test minicluster-snapshot-test --gtest_filter Colocation/MasterExportSnapshotTest.ExportSnapshotAsOfTimeWithHiddenTables/1

Reviewers: asrivastava

Reviewed By: asrivastava

Subscribers: ybase, bogdan, slingam

Differential Revision: https://phorge.dev.yugabyte.com/D33536