[#21651] YSQL: Handle RPC errors as warning while destroying virtual wal
Summary:
Once the LogicalReplication ends either when a client ends it or due to an error, the walsender tries to cleanup the virtual wal instance associated with it.

If there is any error in this cleanup logic, it is better to log it as a warning. This is because:
1. We already have a backup to clean the virtual wal instance
2. The error can mask an actual error due to which the cleanup was required. https://github.com/yugabyte/yugabyte-db/issues/21651 is one such instance.

This revision updates Walsender, to treat any errors received during the DestroyVirtualWALForCDC as a warning.
Jira: DB-10545

Test Plan:
Jenkins: test regex: .*ReplicationSlot.*

ybd --java-test 'org.yb.pgsql.TestPgReplicationSlot#testWalsenderGracefulShutdownWithCDCServiceError'

Reviewers: asrinivasan

Reviewed By: asrinivasan

Subscribers: ycdcxcluster, yql

Differential Revision: https://phorge.dev.yugabyte.com/D33538