[BACKPORT 2024.1][#21530][#21400][#21651] YSQL: Backport walsender fixes for PG compatible logical replication support
Summary:
##### Backport Description
No merge conflicts.

##### Original Description
Original commits:
1604f1bbaf132d7cdf647144687dd64447b8443a / D33432
5c023f3262f1be24d574da22f90e5fbd3c5e06c5 / D33156
2327b8dcfd3a4a4e0382f7ef275342a67920e214 / D33538

###### [#21530] YSQL: Store cached record batch in a separate context

The walsender while streaming records to the client, fetches a batch of records from the CDC services and streams them one by one. Once a
batch is streamed completely, it is freed and the next batch is fetched. A record batch contains the pg datum values representing the DML column
values. They are not that easy to free.

An easier way is to store them in a separate memory context and free it whenever needed.

This revision does that. We now store the record batch in a separate memory context and reset it before making the next call to the CDC service.
Jira: DB-10412

###### [#21400] YSQL: Handle START_REPLICATION with start_lsn > confirmed_flush_lsn

In the `START_REPLICATION` command, the client can specify the LSN from which they want the records to be streamed from.

This revision introduces support for cases when start_lsn > confirmed_flush_lsn. Before this, the value of start_lsn was being ignored and all records from confirmed_flush_lsn onwards were streamed in case of restarts.

Now, the records which belong to a transaction whose commit_lsn < start_lsn are skipped in the decoding phase (yb_decode.c) similar to how it is done in vanilla PG
(decode.c).
Jira: DB-10292

###### [#21651] YSQL: Handle RPC errors as warning while destroying virtual wal

Once the LogicalReplication ends either when a client ends it or due to an error, the walsender tries to cleanup the virtual wal instance associated with it.

If there is any error in this cleanup logic, it is better to log it as a warning. This is because:
1. We already have a backup to clean the virtual wal instance
2. The error can mask an actual error due to which the cleanup was required. https://github.com/yugabyte/yugabyte-db/issues/21651 is one such instance.

This revision updates Walsender, to treat any errors received during the DestroyVirtualWALForCDC as a warning.
Jira: DB-10545

Test Plan:
Jenkins: test regex: .*ReplicationSlot.*

Local load testing. No behavioural changes

./yb_build.sh --java-test 'org.yb.pgsql.TestPgReplicationSlot#testStartLsnValues'
./yb_build.sh --java-test 'org.yb.pgsql.TestPgReplicationSlot#testWalsenderGracefulShutdownWithCDCServiceError'

Reviewers: asrinivasan, skumar, siddharth.shah

Reviewed By: siddharth.shah

Subscribers: yql, ycdcxcluster

Tags: #jenkins-ready

Differential Revision: https://phorge.dev.yugabyte.com/D34163