[#20726] YSQL: Introduce support for streaming update operation via Walsender
Summary:
Before this revision, we only supported streaming Insert & Delete operations as part of the PG compatible logical replication support. This were added in
https://phorge.dev.yugabyte.com/D31997 and https://phorge.dev.yugabyte.com/D32868 respectively.

This revision extends the support to also update operations. To do this, the following has been implemented:
1. Decoding update operations and streaming them
2. Ability to get the replica identities stored in the stream metadata from yb-master and use it while decoding update records. It is used to determine whether we want to handle omitted values differently vs NULL values. We only want to do that for YB specific replica identities.

---

A key part of this support requires the capability to distinguish between NULL and omitted values for YB specific replica identity values. For example: consider replica identity (record type) `CHANGE`:

```
create table t (a int primary key, b int, c text);

insert into t values (1,1, 'abc');

update t set b = 2 where a = 1;      // Statement 1
update t set b = 3, c = NULL where a = 1;       // Statement 2
```

In Statement 1, c is not being updated while in statement 2, it is being explicit set to `NULL`. For statement 1, we need to be able to tell the client that we have decided not to send the column value for `c` as it is unchanged. We cannot send `c` as NULL as that would imply that the value was explicitly set to `NULL`. To do this, we will send omitted values as `'u'` which indicates an unchanges toasted value. The PG debezium connector (our fork) will be modified to treat this as a special case.

This requirement doesn't arise in replica identity values inherited from PG (DEFAULT, NOTHING, FULL) because you can deduce whether the column is omitted or NULL just through the replica identity value.

For e.g., consider the same example as above but with `FULL`. In this case, you'll also get the complete old tuple i.e. none of the columns are omitted. So if a column is sent as NULL, it means that it truly was NULL.

Similar deductions apply to `DEFAULT` and `NOTHING` which have no before image except primary keys.

---

**Upgrade/Rollback safety:**
The logical replication support is disabled via the test flag `ysql_TEST_enable_replication_slot_consumption`.
Jira: DB-9729

Test Plan:
Jenkins: test regex: .*CDCSDK.*

`./yb_build.sh --java-test 'org.yb.pgsql.TestPgReplicationSlot'`

Reviewers: aagrawal, asrinivasan, skumar

Reviewed By: asrinivasan

Subscribers: hsunder, ybase, ycdcxcluster, yql, bogdan

Differential Revision: https://phorge.dev.yugabyte.com/D32859