[#20234] DocDB: Stage 1.1: Removed capabilities from the entire code base
Summary:
Previously, our system used a feature called "capabilities" to track what each server in our distributed database could do. However, since we've introduced a more dynamic way of handling these capabilities through "auto flags," the old capabilities system has become outdated and unnecessary.

In our ongoing project to improve the meta-cache (which helps our system efficiently locate which server is currently leading for a specific piece of data), we've encountered a problem. The meta-cache still holds a list of these outdated capabilities for each server, but we no longer have a straightforward way to update or use this capabilities information effectively. This is because the capabilities of the servers involved in the consensus process (raft peers) can't be easily determined by querying a single server.

To solve this issue, this change removes the outdated capabilities information from the entire codebase. We're also removing any tests that were specifically checking for this capabilities data. By doing this, we streamline the meta-cache structure and remove unnecessary complexity, laying better groundwork for further improvements in how we track and update leadership information in our distributed system.

This simplification helps us focus on more efficient and up-to-date methods of ensuring our database operations are directed to the right servers, enhancing overall performance and reliability.

**Upgrade/Rollback safety:**
The last capability was added in 2.6, and all customers are on version higher than this, so it is safe to remove this feature.
Jira: DB-9194

Test Plan: Build and test with Jenkins

Reviewers: mlillibridge, hsunder, sergei

Reviewed By: mlillibridge, hsunder, sergei

Subscribers: bogdan, yql, ycdcxcluster, ybase

Differential Revision: https://phorge.dev.yugabyte.com/D33598