[PLAT-13325] Retrieve the provider object from the DB during imageBundleSetup
Summary:
For YBM, we use v1 calls for configuring the provider. During this process, we configure the sshPort during the bootstrap call, which is set in the provider details during AccessKeySetup.
CloudTaskBase returns the cached provider which still contains the stale port, thereby causing the bundle to be generated with the wrong port. Adds the capability to forceFetch the provider object from the DB & use the same during bundle setup.

Test Plan:
Created a provider using v1 payload.
Verified that the ports are correctly configured for the bundle

Reviewers: vbansal

Reviewed By: vbansal

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D33612