[#21723] CDCSDK: Fixed segementation fault observed in tablet split tests with replica identity
Summary:
The method `CheckBeforeImageActive()`, (which is called from UpdatePeersAndMetrics) gets
the tablet peer using the tablet_id from `tablet_map_`. However if the tablet is split, then there
is a possibility of parent tablet entry being deleted from the `tablet_map_` but not from the
`cdc_state` table. In such a scenario the tablet peer obtained is null. Trying to access this
resulted in a segmentation fault.

As a fix, we would simply log a warning on finding a null tablet peer and proceed without updating
the peer in the particular iteration of UpdatePeersAndMetrics. In the subsequent iterations, the
state table entry for the parent tablet will also be removed and CheckBeforeImageActive will not try
to get the deleted tablet's peer.
Jira: DB-10597

Test Plan: Jenkins: test regex: .*CDCSDK.*

Reviewers: asrinivasan, stiwary, siddharth.shah

Reviewed By: asrinivasan

Subscribers: ycdcxcluster

Tags: #jenkins-ready

Differential Revision: https://phorge.dev.yugabyte.com/D33617