[BACKPORT 2.20.3][PLAT-13308] Do not ignore index tables when computing tables for bootstrap params
Summary:
When adding a table requiring bootstrap to an xCluster config which already has an index table in replication, the YBA UI does not send the existing index table as part of the bootstrap params.
This is an issue because we're passing `autoIncludeIndexTables = false` to the edit xCluster/DR endpoints which means the YBA UI is responsible for making sure the index table is present in the bootstrap params.

This diff ensures we're not ignoring the index tables when computing the set of tables for bootstrap params.

This diff also removes the ability to add/remove index tables from the edit table modal for transactional xCluster configs.
Adding index tables to transactional xCluster configs is done by the DB. Users need to update the xCluster config on YBA by running the reconcile replication action.
Users can remove index tables by removing them individually on the table list.

Original commit: 90d6f7efed3c23fb85ffd931ea484120f4e2b4f8 / D33577

Test Plan:
**Test Matrix**
{F165356}

Reviewers: sanketh, cwang, rmadhavan

Reviewed By: rmadhavan

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D33606