[#21701] Send the replica identity of the table which existed at the time of slot creation.
Summary:
As part of the RELATION message, the replica identity of the relation is sent during logical replication.

In YSQL, the effective replica identity that is used is the value which existed at the time of slot creation. Future Alter Table Replica Identity statements only
affect streams created after it but not before it. As a result, we now send the correct effective replica identity as part of the RELATION message.

Test Plan:
Jenkins: test regex: .*ReplicationSlot.*

./yb_build.sh --java-test 'org.yb.pgsql.TestPgReplicationSlot#testWithAlterReplicaIdentity'

Reviewers: asrinivasan

Reviewed By: asrinivasan

Subscribers: yql, ycdcxcluster

Differential Revision: https://phorge.dev.yugabyte.com/D33649