Use `to_regtypemod` on Postgres 17
When running on Postgres 17 or higher, take advantage of the new
`to_regtypemod` function to much more correctly and efficiently
normalize type names, including their type modifiers. Notably, interval
types will now also be normalized, so note this fact in the docs.

Update the CI testing to test on Postgres 17 and adjust patch offsets
for the line count change in `pgtap.sql.in`.