[#21773] YSQL: Fix rpc callback invoked on deleted ddl bg task
Summary:
The relevant code that can lead to rpc callback invocation on a deleted ddl
background task is in `PollTransactionStatusBase::VerifyTransaction()`

```
  *rpc_handle = client::GetTransactionStatus(
    TransactionRpcDeadline(),
    nullptr /* tablet */,
    client,
    &req,
    [this, rpc_handle]
        (Status status, const tserver::GetTransactionStatusResponsePB& resp) {
      auto retained = rpcs_.Unregister(rpc_handle);
      TransactionReceived(std::move(status), resp);
    });
```

The base class `PollTransactionStatusBase` has two subclasses:

```
class NamespaceVerificationTask : public MultiStepNamespaceTaskBase,
                                  public PollTransactionStatusBase {
```
and
```
class TableSchemaVerificationTask : public MultiStepTableTaskBase,
                                    public PollTransactionStatusBase {
```

Both tasks can be asynchronously deleted by `TasksTracker::CleanupOldTasks`.
Therefore it is possible that the background task is already deleted before the
callback shown above is invoked. This can lead to yb-master crash.

To fix this bug, we need to ensure that the bg task isn't deleted before the
callback is invoked. I added a new virtual function `GetSharedFromThis()` to
retrieve a `shared_ptr` to the bg task object and capture it within the callback
itself as a weak_ptr<MonitoredTask>.
Jira: DB-10649

Test Plan: jenkins

Reviewers: hsunder, tverona

Reviewed By: hsunder

Subscribers: ybase, yql, bogdan

Differential Revision: https://phorge.dev.yugabyte.com/D33731