[#20234] DocDB: Stage 1: Added opid index into meta-cache when LookUpById response returns
Summary:
Problem Background:
In our system, when a client needs to perform an operation on a specific tablet, it first needs to find out which server is currently responsible for that operation. If the operation is a WriteRpc for example, it must find the tablet leader server. However, the system's current method of figuring out the tablet leader is not very efficient. It tries to guess the leader based on a list of potential servers (peers), but this guessing game can be slow, especially when there are many servers or when the servers are located far apart geographically. This inefficiency can lead to operations failing because the leader wasn't found quickly enough.

Additionally, the system doesn't handle server failures well. If a server is down, it might take a long time for the system to stop trying to connect to it, wasting valuable seconds on each attempt. While there's a mechanism to avoid retrying a failed server for 60 seconds, it's not very effective when a server is permanently out of service. One reason for this inefficiency is that the system's information about who the leaders are (stored in something called the meta cache) can become outdated, and it doesn't get updated if the system can still perform its tasks with the outdated information, even if doing so results in repeated connection failures.

Solution Introduction:
This ticket introduces a preliminary change aimed at improving how the system tracks the current leader for each piece of data. The idea is to add a new piece of information to the meta cache called "raft_config_opid," which records the latest confirmed leadership configuration for each tablet. This way, when the system receives new information about the leadership configuration (which can happen during normal operations from other servers), it can check this new information against what it already knows. If the new information is more up-to-date, the system can update its meta cache, potentially avoiding wasted efforts on trying to connect to servers that are no longer leaders or are down.

In simpler terms, the goal is to make the system smarter about who it tries to talk to when performing operations, which should help it avoid unnecessary delays and failures caused by outdated information or attempts to communicate with servers that aren't currently leading or are unavailable. This change sets the stage for further improvements that will rely on the new "raft_config_opid_index" information to keep the meta cache up-to-date.

**Upgrade/Rollback safety:**
The added field in the response is not to be persisted on disk, it is guarded by protobuf's backward compatibility
Jira: DB-9194

Test Plan: Manual testing by inspecting meta-cache endpoint, verify that the raft_config_opId_index field is actually being updated

Reviewers: mlillibridge, arybochkin

Reviewed By: mlillibridge, arybochkin

Subscribers: hsunder, arybochkin, ybase, bogdan

Differential Revision: https://phorge.dev.yugabyte.com/D33197