[#21765] YSQL: Fix handling of serialized transactions in reorder buffer
Summary:
During logical replication via the walsender, if a transaction has more than max_changes_in_memory (4096) DML records, it is serialized to the disk. They are
written to the `pg_repl_slot/{slot_name}/` directory. In PG, this directory is created as part of the slot creation and also stores the stream metadata and any
exported snapshots.

In YB, we had disabled the creation of this directory during slot creation because:
1. yb-master is the source of truth
2. One could create the slot on one node and then connect on another to consume changes.

This meant that when a transaction needs to be serialized, the reorder buffer would run into error since the directory doesn't exist.

This revision add the logic to create this directory upon the start of replication. This needs to happen at the start of streaming instead of slot creation due to the
point 2 above.
Jira: DB-10640

Test Plan:
Jenkins: test regex: .*ReplicationSlot.*

```
./yb_build.sh --java-test 'org.yb.pgsql.TestPgReplicationSlot#testReplicationWithSerializedTransaction'
```

Reviewers: asrinivasan

Reviewed By: asrinivasan

Subscribers: ycdcxcluster, yql

Differential Revision: https://phorge.dev.yugabyte.com/D33750