[#29018] YSQL: Deprecate enable_pg_savepoints
Summary:
The tserver gflag `enable_pg_savepoints` is being deprecated because of 2 reasons:

(1) Read committed isolation registers an internal savepoint before each new statement
in the transaction. This is required to be able to rollback any work when retrying
a statement on kConflict and kReadRestart errors. If savepoints are disabled, RC
transactions would silently result in incorrect behaviour. The plan was to add an
enforcement to FATAL if `enable_pg_savepoints=false` when
`yb_enable_read_committed_isolation=true`. However, it is better to get rid of
`enable_pg_savepoints` instead of the enforcement because this flag is not
really used anywhere and the savepoints feature has hardened over time. We
don't need to give the user an option to disable it.

(2) Even when a user sets `enable_pg_savepoints=false`, pl/pgsql exceptions
which internally use savepoints don't error out, instead they silently make
progress without really creating sub-txns. This is not a good behaviour to keep.

Test Plan: Jenkins

Reviewers: patnaik.balivada, rsami, mlillibridge

Reviewed By: patnaik.balivada

Subscribers: yugaware, ybase, yql

Differential Revision: https://phorge.dev.yugabyte.com/D33785