[#21544] xCluster: Support colocated DBs in DB Scoped replication
Summary:
xCluster setup requires at least one table to be present in the database.
For colocated databases we need to have the colocated table group, and at least one colocated table. This is because we map the DocDB table Ids of the tablegroups, between the two universes using the tables in the groups.
This change adds checks to make sure these conditions are satisfied.

- Moved `IsSequencesSystemTable` from `CatalogManager` to `TableInfo`
- Created a common `IsTableEligibleForXClusterReplication` util function that can be used for both source and target side checks.

Fixes #21544
Jira: DB-10429

Test Plan:
XClusterDBScopedTest.TestCreateWithCheckpoint
XClusterDBScopedTest.ColocatedDB

Reviewers: jhe, slingam, xCluster

Reviewed By: jhe

Subscribers: yql, xCluster, ybase, bogdan

Differential Revision: https://phorge.dev.yugabyte.com/D33688