[pg15-upgrade] feat: enable pg_upgrade with basic functionality
Summary: This revision enables pg_upgrade binary with basic functionality to dump and restore database metadata. Currently, it only migrates the metadata for the 'yugabyte' database, other databases will be enabled later. There is also a modification to pg_upgrade to take host addresses as it did not have that option. The preflight checks for checking cluster compatibility are currently disabled.  The test_upgrade is modified to use pg_upgrade binary instead of executing ysql_dump and pg_restore separately.

Test Plan:
```
./yb_build --sj release
./pg15_tests/test_upgrade.sh
./pg15_tests/test_upgrade_comment.sh
./pg15_tests/test_upgrade_dropped_col.sh
```

Reviewers: tfoucher

Reviewed By: tfoucher

Subscribers: yql, tfoucher, jason

Differential Revision: https://phorge.dev.yugabyte.com/D33848